+++ test/integration/thermal_camera.cc
@@ -32,6 +32,7 @@
 #include <gz/common/Event.hh>
 #include <gz/sensors/Manager.hh>
 #include <gz/sensors/ThermalCameraSensor.hh>
+#include <gz/utils/ExtraTestMacros.hh>
 
 // TODO(louise) Remove these pragmas once gz-rendering is disabling the
 // warnings
@@ -357,7 +358,9 @@ void ThermalCameraSensorTest::ImagesWithBuiltinSDF(
 }
 
 //////////////////////////////////////////////////
-TEST_P(ThermalCameraSensorTest, ImagesWithBuiltinSDF)
+// See: https://github.com/gazebosim/gz-rendering/issues/654
+TEST_P(ThermalCameraSensorTest,
+       GZ_UTILS_TEST_DISABLED_ON_MAC(ImagesWithBuiltinSDF))
 {
   ImagesWithBuiltinSDF(GetParam());
 }