+++ Changelog.md
@@ -6,6 +6,16 @@
 
 ## Gazebo Sensors 6
 
+### Gazebo Sensors 6.6.0 (2022-06-17)
+
+1. Add BoundingBox Sensor
+    * [Pull request #136](https://github.com/gazebosim/gz-sensors/pull/136)
+
+### Gazebo Sensors 6.5.0 (2022-05-24)
+
+1. Add HasConnections function
+    * [Pull request #222](https://github.com/gazebosim/gz-sensors/pull/222)
+
 ### Gazebo Sensors 6.4.0 (2022-05-13)
 
 1. Set lidar visibility mask